Solve warnings on new Github virtual environments (#333)
* Fix out of bounds error for 2D turbinflow.

* Use nested namespaces.

* Fix another 2D out of bounds.

* Explicit casts.

* Fix more implicit casts.

* Use ubuntu-20.04 for matrix.

+++ Reactions/ReactorArkode.cpp
@@ -146,7 +146,7 @@ ReactorArkode::react(
 {
   BL_PROFILE("Pele::ReactorArkode::react()");
 
-  const int ncells = box.numPts();
+  const int ncells = static_cast<int>(box.numPts());
   AMREX_ASSERT(ncells < std::numeric_limits<int>::max());
 
   const int neq = NUM_SPECIES + 1;
