+++ Reactions/ReactorBase.cpp
@@ -27,7 +27,7 @@ namespace pele::physics::reactions {
 void
 ReactorBase::set_typ_vals_ode(const std::vector<amrex::Real>& ExtTypVals)
 {
-  const int size_ETV = ExtTypVals.size();
+  const int size_ETV = static_cast<int>(ExtTypVals.size());
   amrex::Vector<std::string> kname;
   pele::physics::eos::speciesNames<pele::physics::PhysicsType::eos_type>(kname);
   int omp_thread = 0;
